requirements.txt

flask
charm-crypto
